pyproject.toml
[project]
name = "fastfollow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=3.0.3",
    "gradio>=5.46.0",
    "python-dotenv>=1.1.1",
]
